User age without cast

+++ src/main/java/jm/task/core/jdbc/Main.java
@@ -2,7 +2,6 @@ package jm.task.core.jdbc;
 
 import jm.task.core.jdbc.dao.UserDaoJDBCImpl;
 import jm.task.core.jdbc.model.User;
-import java.sql.*;
 import java.util.List;
 
 public class Main {
@@ -10,10 +9,10 @@ public class Main {
         List<User> users;
         UserDaoJDBCImpl userDao = new UserDaoJDBCImpl();
         userDao.createUsersTable();
-        userDao.saveUser("Александр", "Иванов", (byte) 30);
-        userDao.saveUser("Сергей", "Петров", (byte) 29);
-        userDao.saveUser("Олег", "Сидоров", (byte) 40);
-        userDao.saveUser("Ольга", "Иванова", (byte) 28);
+        userDao.saveUser("Александр", "Иванов",30);
+        userDao.saveUser("Сергей", "Петров", 29);
+        userDao.saveUser("Олег", "Сидоров", 40);
+        userDao.saveUser("Ольга", "Иванова", 28);
         users = userDao.getAllUsers();
         System.out.println(users);
         userDao.cleanUsersTable();
